backend/requirements.txt
flask>=3.0
flask-cors>=4.0
pandas>=2.0
openpyxl>=3.1
numpy>=1.26
scikit-learn>=1.4
joblib>=1.3
PyMuPDF>=1.23
python-docx>=1.1
